feat(mindshare lead popup): mindshare popups
