+++ src/views/zesty/Article.js
@@ -110,6 +110,18 @@ function Article({ content }) {
     );
   }, []);
 
+  const popupLeadCaptureProps = {
+    title: content?.pop_up_title || 'FREE CMS BUYING GUIDE',
+    description: content?.pop_up_description || 'DOWNLOAD CMS BUYING GUIDE',
+    ctaText: content?.pop_up_cta_text || 'DOWNLOAD NOW',
+    thankYouMessage:
+      content?.pop_up_thank_you ||
+      'Thank you for downloading our CMS buying guide. Reach out to us for a free discovery call, demo call or a free trial',
+    pdfLink:
+      content?.pdf_link ||
+      'https://kfg6bckb.media.zestyio.com/HeadlessCMS-Buyers-Guide-Zesty.H17lCRwtp.pdf',
+  };
+
   return (
     <Box sx={{ position: 'relative' }}>
       <ThemeProvider theme={() => revampTheme(palette.mode)}>
@@ -607,7 +619,7 @@ function Article({ content }) {
       )}
 
       {/* Side PopUp */}
-      <PopUpLeadCapture />
+      <PopUpLeadCapture {...popupLeadCaptureProps} />
     </Box>
   );
 }
